Stabilize executable path parsing
If the .exe extension is not present in the command line, the definitive source of truth for the exe is the image name parameter. When the process image is loaded for the first time, we'll check the length of the current exe, and if lower than the one given in the load image param, we'll use the latter.