Revamp rule actions
The rule action block defined
as a Go template is very unintuitive and hard to grok for users not familiar with the language. Generating an alert should be a default behaviour for all rules. This is why we shift from explicit to implicit alert generation. The action block is now an array of predefined response actions, such as `kill`. When the alert is generated, the rule name is used as its title. For building a more informative alert, the text can be built via the new `output` element, which accepts filter fields expressions.