Remove kill action pid field
Introducing a field to specify which
process id is used to terminate the
process may be convoluted and at the same time superfluous to handle. For this reason, the pid field is removed from the action definition, and the action automatically collects pids to be killed by walking event matches.