Revert system providers
Setting up separate tracing sessions is
not viable as the events arrive out of
order and some fundamental events, such as
CreateProcess are emitted after other system
events leading to empty process state. For this reason, the only provider that is run
in its own tracing session is the registry provider as it can generate a massive amount of events.