Handle unexported symbols
When resolving the symbol name from the
export directory, if the function RVA is
outside export directory range or the export directory is empty, we don't
fallback to Debug Help API for symbol
resolution, as it wouldn't be able to
retrieve symbol information.

Additionally, the PE exports now also
contain ordinal symbols.