Add new rule attributes
A new set of attributes is added to describe
characteristics of the specific rule. Firstly,
a unique identifier is assigned to the rule. This permits tracking the rule even if its name or condition changes. Next, the rule version attribute is added. Both of these
attributes are required. Additionally, references and notes attributes describe various resources such as blog posts or detection guidelines. Lastly, the notes attribute allows planting any content that
further describes the rule or leaves comments
for security analysts.