refactor(alertsenders/mail): Move template rendering to sender
HTML templates are more likely to exclusively be used
by the mail sender. For this purpose, a new config flag
is defined that determines if the alert is sent with
HTML template or plain text. Alert structure is augmented
with labels and description fields to make it compatible
with the same fields available in the filter config.