feat(yara): Revamp Yara scanner
The Yara scanner is revamped to perform file and
memory scanning triggered by multiple signals. Aside
from the basic process creation and image loading, the
scan is initiated when the PE file is dropped in the file
system, or when the ADS (Alternate Data Stream) is created.
Memory scan is triggered under suspicious memory allocation or section mapping. Lastly, when the
registry binary value is set, the scan is also performed
on the binary blob.