refactor(alertsenders): Launch systray server manually
The systray server is automatically launched when the
systray alert sender is enabled. The process is created
either via the regular CreateProcess API or CreateProcessAsUser
when Fibratus is running inside Windows Service.
This requires querying the user token of the currently logged-in
console session user. The TCB privilege must be present in the
user token.