feat(instrumentation,telemetry): Incorporate thread pool telemetry
The thread pool telemetry is captured from the
dedicated thread pool provider. Presently, we
collect three different event types including
the call stacks. More specifically, SubmitThreadpoolWork, SubmitThreadpoolCallback, and
SetThreadpoolTimer events.
Additionally, the callback addresses are symbolized
to derive to function name, and when the callback is
the ZwContinue or RtlCaputreContext function call, we
also try to decode the CONTEXT structure and symbolize
the instruction pointer address.