fix(rules): Use the correct form of the not operator
The not operator has two variants. If used in front of
the operator it negates the value of the operator eval.
If used in front of the binary expression, it negates
the entire expression. If not used adequately, this can
introduce issues. So, it is better to standardize the
usage of the not operator and if meant to solely negate
the result of the operator use the lhs not operator rhs
pattern.