refactor: Trim the `k` prefix from types and packages
Historically, Fibratus has been envisioned as a tool to exclusively
interact with the NT Kernel Logger ETW provider. Since then, more
providers have been integrated, some of them operating in userspace.
Thus, the assumption that the provenance of all events is coming out
the kernel is no longer valid. It is semantically more correct to represent
all types, packages, and identifier in a generic way. The most prominent
example is the `Kevent` structure being renamed to `Event`.